+++ feature/recruitment/src/main/java/in/koreatech/koin/feature/recruitment/ui/main/RecruitmentMainScreen.kt
@@ -15,7 +15,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.CircularProgressIndicator
@@ -27,9 +26,7 @@ import androidx.compose.material3.pulltorefresh.PullToRefreshBox
 import androidx.compose.material3.pulltorefresh.PullToRefreshDefaults.Indicator
 import androidx.compose.material3.pulltorefresh.rememberPullToRefreshState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -50,6 +47,7 @@ import `in`.koreatech.koin.feature.recruitment.R
 import `in`.koreatech.koin.feature.recruitment.model.RecruitmentCategory
 import `in`.koreatech.koin.feature.recruitment.model.RecruitmentLocation
 import `in`.koreatech.koin.feature.recruitment.model.RecruitmentStatus
+import `in`.koreatech.koin.feature.recruitment.ui.component.rememberRecruitmentPaginationListState
 import `in`.koreatech.koin.feature.recruitment.ui.main.component.RecruitmentAppliedFilterChipGroup
 import `in`.koreatech.koin.feature.recruitment.ui.main.component.RecruitmentChip
 import `in`.koreatech.koin.feature.recruitment.ui.main.component.RecruitmentChipDefaults
@@ -60,12 +58,8 @@ import `in`.koreatech.koin.feature.recruitment.ui.main.model.RecruitmentFilterSt
 import `in`.koreatech.koin.feature.recruitment.ui.main.model.RecruitmentItemModel
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.filter
 import org.orbitmvi.orbit.compose.collectAsState
 import org.orbitmvi.orbit.compose.collectSideEffect
-
-private const val LOAD_MORE_THRESHOLD = 3
 
 @Composable
 fun RecruitmentMainScreen(
@@ -278,18 +272,12 @@ private fun RecruitmentMainScreenImpl(
                         RecruitmentEmptyContent(modifier = Modifier.padding(bottom = 40.dp))
                     }
                 } else {
-                    val listState = rememberLazyListState()
-
-                    LaunchedEffect(listState, hasMore, isLoadingMore, items.size) {
-                        snapshotFlow {
-                            val layoutInfo = listState.layoutInfo
-                            val lastVisibleItemIndex = layoutInfo.visibleItemsInfo.lastOrNull()?.index ?: 0
-                            lastVisibleItemIndex >= layoutInfo.totalItemsCount - LOAD_MORE_THRESHOLD
-                        }
-                            .distinctUntilChanged()
-                            .filter { it }
-                            .collect { if (hasMore && !isLoadingMore) onLoadMore() }
-                    }
+                    val listState = rememberRecruitmentPaginationListState(
+                        hasMore = hasMore,
+                        isLoadingMore = isLoadingMore,
+                        itemCount = items.size,
+                        onLoadMore = onLoadMore
+                    )
 
                     LazyColumn(
                         modifier = Modifier.fillMaxSize(),
