+++ feature/recruitment/src/main/java/in/koreatech/koin/feature/recruitment/ui/myrecruitment/MyRecruitmentScreen.kt
@@ -12,16 +12,13 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
@@ -34,6 +31,7 @@ import `in`.koreatech.koin.feature.recruitment.R
 import `in`.koreatech.koin.feature.recruitment.model.RecruitmentCategory
 import `in`.koreatech.koin.feature.recruitment.model.RecruitmentRole
 import `in`.koreatech.koin.feature.recruitment.ui.component.RecruitmentFilterButton
+import `in`.koreatech.koin.feature.recruitment.ui.component.rememberRecruitmentPaginationListState
 import `in`.koreatech.koin.feature.recruitment.ui.myrecruitment.component.CloseRecruitmentDialog
 import `in`.koreatech.koin.feature.recruitment.ui.myrecruitment.component.MyRecruitmentEmptyState
 import `in`.koreatech.koin.feature.recruitment.ui.myrecruitment.component.RecruitmentFilterBottomSheet
@@ -42,12 +40,8 @@ import `in`.koreatech.koin.feature.recruitment.ui.myrecruitment.model.MyRecruitm
 import `in`.koreatech.koin.feature.recruitment.ui.myrecruitment.model.RecruitmentStatus
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.filter
 import org.orbitmvi.orbit.compose.collectAsState
 import org.orbitmvi.orbit.compose.collectSideEffect
-
-private const val LOAD_MORE_THRESHOLD = 3
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -152,18 +146,12 @@ private fun MyRecruitmentScreenImpl(
                 MyRecruitmentEmptyState()
             }
         } else {
-            val listState = rememberLazyListState()
-
-            LaunchedEffect(listState, hasMore, isLoadingMore, posts.size) {
-                snapshotFlow {
-                    val layoutInfo = listState.layoutInfo
-                    val lastVisibleItemIndex = layoutInfo.visibleItemsInfo.lastOrNull()?.index ?: 0
-                    lastVisibleItemIndex >= layoutInfo.totalItemsCount - LOAD_MORE_THRESHOLD
-                }
-                    .distinctUntilChanged()
-                    .filter { it }
-                    .collect { if (hasMore && !isLoadingMore) onLoadMore() }
-            }
+            val listState = rememberRecruitmentPaginationListState(
+                hasMore = hasMore,
+                isLoadingMore = isLoadingMore,
+                itemCount = posts.size,
+                onLoadMore = onLoadMore
+            )
 
             LazyColumn(
                 modifier = Modifier.fillMaxSize(),
